+++ boilerplates/nuxt/__tests__/Breadcrumbs/BreadcrumbsWithIcon.spec.js
@@ -1,0 +1,12 @@
+import { shallowMount } from '@vue/test-utils';
+import BreadcrumbsWithIcon from '../../components/Breadcrumbs/BreadcrumbsWithIcon.vue';
+let wrapper;
+
+describe('BreadcrumbsWithIcon.vue', () => {
+  beforeEach(() => {
+    wrapper = shallowMount(BreadcrumbsWithIcon);
+  });
+  it('render component correctly', () => {
+    expect(wrapper.html()).toMatchSnapshot();
+  });
+});