+++ boilerplates/nuxt/jest.config.js
@@ -24,4 +24,4 @@ module.exports = {
       lines: 80
     }
   }
-}
+};